requirements.txt
django-allauth>=65.4.1
django-crispy-forms>=2.3
djangorestframework>=3.15.2
django>=5
pymysql
whitenoise
django-admin-sortable>=2.3
django-environ>=0.12.0
django-utils-six>=2.0
django-admin-action-forms>=1.3.0
django-map-widgets>=0.5.1
django-chartjs>=2.3.0
python-pesa>=0.7
djangorestframework
markdown>=3.7
django-filter>=25.1